requirements.txt
sqlalchemy
psycopg2 
python-dotenv
fastapi
uvicorn
alembic